General: Cancun Stop
Calling cancun.sh stop will kill all cancun processes

+++ cancun.sh
@@ -2,12 +2,9 @@
 
 stop() {
     pid=`ps -ef | grep 'cancun.py' | awk '{ print $2 }'`
-    echo $pid
     kill $pid
-    #sleep 2
     echo "Cancun Project Killed!"
-    #test -f /tmp/cancun.pid && rm -f /tmp/cancun.pid
-    rm -f /tmp/cancun.pid
+    test -f /tmp/cancun.pid && rm -f /tmp/cancun.pid
     /home/irlp/bin/forceunkey
 }
 
@@ -16,8 +13,7 @@ case "$1" in
     stop
     ;;
   *)
-    echo "Usage: cancun.sh stop}"
+    echo "Usage: cancun.sh stop"
     exit 1
 esac
 
-exit 0
